# Use non-capture groups for allowlists