### Description:
[Non-capture groups appear to be measurably faster.](https://github.com/leaktk/patterns/pull/59#discussion_r1932773747)

**Benchmark**
Regexes
```
# Before
BenchmarkConfigAllowlistRegexes
BenchmarkConfigAllowlistRegexes-4   	    1972	    519161 ns/op
# After
BenchmarkConfigAllowlistRegexes
BenchmarkConfigAllowlistRegexes-4   	    2899	    390059 ns/op
```

Paths
```
# Before
BenchmarkConfigAllowlistPaths
BenchmarkConfigAllowlistPaths-4   	    1980	    509435 ns/op

# After
BenchmarkConfigAllowlistPaths
BenchmarkConfigAllowlistPaths-4   	    2912	    385768 ns/op
```

### Checklist:

* [x] Does your PR pass tests?
* [x] Have you written new tests for your changes?
* [x] Have you lint your code locally prior to submission?