@bplaxco 2025-01-28
@rgmz nice! glad that resulted in a performance improvement :D