@zricethezav 2025-01-29
This is fantastic. A follow up PR/issue could be to check each allowlist to make sure capture groups are not used during config generation. I _think_ the regex pkg exposes that but I may be wrong